Generate multiple reports from multiple arguments to analyze-clusterbuster-report
