+++ analyze-clusterbuster-report
@@ -16,6 +16,8 @@
 import sys
 import argparse
 import json
+import re
+import os
 from lib.clusterbuster.analysis.ClusterBusterAnalysis import ClusterBusterAnalysis
 from lib.clusterbuster.loader.ClusterBusterLoader import ClusterBusterLoader
 
@@ -24,6 +26,7 @@ analysis_formats = ClusterBusterAnalysis.list_analysis_formats()
 
 parser.add_argument("-o", "--outfile", default=None, type=str, metavar='file', help='Output filename')
 parser.add_argument('--list_formats', action='store_true', help='List available report formats')
+parser.add_argument('--kata', action='store_true', help='Compare Kata with non-Kata results')
 parser.add_argument("-r", "--report-type", default=None, type=str, metavar='format',
                     choices=analysis_formats, help=f'Analysis format: one of {", ".join(analysis_formats)}')
 parser.add_argument("files", metavar='file', type=str, nargs='*', help='Files to process')
@@ -33,7 +36,9 @@ if args.list_formats:
     sys.exit(1)
 
 
-def do_analyze_clusterbuster(f, data):
+def analyze_clusterbuster_1(f, data):
+    json.dump(data, f, indent=2)
+    return
     report = ClusterBusterAnalysis(data, report_type=args.report_type).Analyze()
     if isinstance(report, str):
         print(report, file=f)
@@ -41,13 +46,93 @@ def do_analyze_clusterbuster(f, data):
         json.dump(report, fp=f, indent=2)
 
 
+known_subdirs = {}
+
+
+def matches_patterns(f: str, patterns: list):
+    if patterns:
+        for pattern in patterns:
+            if not re.search(pattern, f):
+                return False
+    return True
+
+
+def create_report_spec(arg: str):
+    answer = {}
+    dirname = None
+    job_patterns = []
+    run_name = None
+    for component in arg.split(':'):
+        pair = component.split('=', 1)
+        if len(pair) == 1:
+            dirname = pair[0]
+        else:
+            key, value = pair
+            key = key.lower()
+            if key.startswith('dir'):
+                dirname = value
+            elif key == 'job_pattern':
+                job_patterns.append(value)
+            elif key == 'name':
+                run_name = value
+            else:
+                raise ValueError(f"Unexpected key {key} in name {arg}")
+    if dirname is None:
+        raise ValueError(f"No directory name found in {arg}")
+    if run_name is None:
+        run_name = dirname.rstrip('/').split('/')[-1]
+    if job_patterns:
+        for pattern in job_patterns:
+            run_name += f'-{pattern.lstrip(" -_").rstrip(" -_")}'
+    if os.path.isdir(dirname):
+        files = [os.path.join(dirname, f) for f in os.listdir(dirname)]
+        dirs = []
+        dirs = [os.path.realpath(d) for d in files if (matches_patterns(d, job_patterns) and
+                                                       os.path.isdir(d) and
+                                                       os.path.isfile(os.path.join(d, "clusterbuster-report.json")))]
+        if not dirs:
+            print(f"No matching subdirectories found under {dirname} ({arg})", sys.stderr)
+            return None
+        for d in dirs:
+            if d in known_subdirs:
+                raise ValueError(f"Directory {d} already used in report {known_subdirs[d]}")
+            else:
+                known_subdirs[d] = arg
+        answer['dirs'] = dirs
+        answer['run_name'] = run_name
+    else:
+        raise ValueError(f"{dirname}: Not a directory")
+    return answer
+
+
+def analyze_clusterbuster(f):
+    reports = {}
+    dir_args = []
+    if args.kata:
+        for arg in args.files:
+            dir_args.append(f'{arg}:job_pattern=-runc-')
+            dir_args.append(f'{arg}:job_pattern=-kata-')
+    else:
+        dir_args = args.files
+    for arg in dir_args:
+        spec = create_report_spec(arg)
+        if spec is not None:
+            if spec['run_name'] in reports:
+                raise ValueError(f'Duplicate report name {spec["run_name"]}')
+            reports[spec['run_name']] = spec
+    if not reports:
+        raise ValueError("No reports found")
+    for report in reports.values():
+        analyze_clusterbuster_1(f, ClusterBusterLoader(report['dirs']).Load())
+
+
+print(args.files, file=sys.stderr)
 try:
-    data = ClusterBusterLoader(args.files).Load()
     if args.outfile and args.outfile != '-':
         with open(args.outfile, 'w') as f:
-            do_analyze_clusterbuster(f, data)
+            analyze_clusterbuster(f)
     else:
-        do_analyze_clusterbuster(sys.stdout, data)
+        analyze_clusterbuster(sys.stdout)
 except KeyboardInterrupt:
     sys.exit(1)
 except BrokenPipeError:
